Add weibo comments node , optimize code structure
