+++ weibo_base/__init__.py
@@ -6,5 +6,9 @@
  File: __init__.py.py
  Time: 5/19/18
 """
+from .weibo_typing import *
 from .weibo_api import *
+from .weibo_component import *
 from .weibo_util import *
+from .weibo_parser import *
+
