+++ weibo_base/weibo_typing.py
@@ -1,0 +1,15 @@
+# -*- coding:utf-8 -*-
+
+"""
+ Verion: 1.0
+ Author: Helixcs
+ Site: https://iliangqunru.bitcron.com/
+ File: weibo_typing.py
+ Time: 11/25/18
+"""
+
+from typing import Optional
+
+_JSONResponse = Optional[dict]
+_StrFieldResponse = Optional[str]
+_IntFieldResponse = Optional[int]
